requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
mysqlclient==2.0.3
bcrypt==3.2.0
pyjwt==2.1.0
